spam time-border
+ add time-borders to all queries

The JabberSpam manifesto states that the critical period for a server to be approved again is 14 days. Thus I adjusted all queries required for the report function to only count within these borders.

Though if a domain is specified without the report argument the time border is increased to 1 month to increase visibility.

misc:
+ add Codefactor badge