feature: custom timeperiod
+ add ability to define a custom time period ( beta )

This feature enables users to query in a custom timeperiod to further specify their results.
Every script feature is able to handle the custom timestamps properly. But beware that timezones different from your local timezone might omit data from your query.
Custom timestamps have to be in the ISO-8601 format, the easiest way is to use the date function to create proper timestamps.

For example:
	$date -Ins -d "3 weeks ago"
	$date --iso-8601 -d "3 weeks ago"